% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_browser_data.R
\name{available_treeview}
\alias{available_treeview}
\title{function to return treeview options}
\usage{
available_treeview(data_dir)
}
\arguments{
\item{data_dir}{The directory containing the data for the application.}
}
\description{
function to return treeview options
}
