% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_browser_data.R
\name{create_all_node_lookups}
\alias{create_all_node_lookups}
\title{Function to create lookups for nodes for all .rds files in the treeview directory}
\usage{
create_all_node_lookups(data_dir)
}
\arguments{
\item{data_dir}{The directory where the data should be read from / written to. This must
contain a `treeview` subdirectory.}
}
\description{
Function to create lookups for nodes for all .rds files in the treeview directory
}
