% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_browser_data.R
\name{create_browser_data}
\alias{create_browser_data}
\title{Generate treeview illustrations and all metadata for presentation by tfpbrowser}
\usage{
create_browser_data(e0, output_dir, ...)
}
\arguments{
\item{e0}{Path to the scanner environment produced by \code{tfpscan}. Alternatively can pass
the environment directly.}

\item{output_dir}{Where should the output files be saved to? This function will create
subdirectories `<output_dir>/treeview`, `<output_dir>/mutations`, `<output_dir>/sequences` and
assumes that `<output_dir>/scanner_output` already exists.}

\item{...}{Further arguments for `treeview()`.}
}
\description{
Generate treeview illustrations and all metadata for presentation by tfpbrowser
}
