% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_sina.R
\name{create_cluster_sina_ggplot}
\alias{create_cluster_sina_ggplot}
\title{Creates an interactive sina plot using ggplot2 and ggiraph}
\usage{
create_cluster_sina_ggplot(
  sc1,
  x_lab = "Lineage and/or mutation",
  y_lab = "logistic_growth_rate"
)
}
\arguments{
\item{sc1}{Data-frame. As created using \code{format_cluster_sina_data}.}

\item{x_lab, y_lab}{Scalar string. Labels for the x- and y-axis.}
}
\value{
an interactive ggplot2 object
}
\description{
Creates an interactive sina plot using ggplot2 and ggiraph
}
