% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_browser_data.R
\name{create_dir_if_missing}
\alias{create_dir_if_missing}
\title{Create a directory if it doesn't yet exist}
\usage{
create_dir_if_missing(path)
}
\arguments{
\item{path}{The path for the required directory}
}
\description{
Create a directory if it doesn't yet exist
}
