% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{create_interactive_ggtree}
\alias{create_interactive_ggtree}
\title{Converts a \code{ggtree} object into a \code{ggiraph} object with interactive potential}
\usage{
create_interactive_ggtree(
  ggobj,
  branch_col,
  cluster_size_range,
  shapes,
  colours,
  colour_limits
)
}
\arguments{
\item{ggobj}{A \code{ggtree} object.}

\item{branch_col}{Scalar string. Name of the column in \code{ggobj$data} that we are
creating the plot for.}

\item{cluster_size_range}{Numeric (length-2). min and max values for cluster sizes on the
chart.}

\item{shapes}{Shapes for the branches and leaves in the tree.}

\item{colours}{Vector of colours.}

\item{colour_limits}{Min and max values for the colours.}
}
\value{
A  \code{ggtree} object with interactive data for presentation by
  \code{ggiraph::girafe}.
}
\description{
Converts a \code{ggtree} object into a \code{ggiraph} object with interactive potential
}
