% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_browser_data.R
\name{create_mutation_files}
\alias{create_mutation_files}
\title{Create .csvs containing defining- and all-mutations}
\usage{
create_mutation_files(e0, mutations_dir)
}
\arguments{
\item{e0}{Path to the scanner environment produced by \code{tfpscan}. Alternatively can pass
the environment directly.}

\item{mutations_dir}{The directory where the mutations .csv files will be placed. The
filenames are `<mutations_dir>/all_mutations.csv` and `<mutations_dir>/defining_mutations.csv`.}
}
\description{
Create .csvs containing defining- and all-mutations
}
