% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{create_noninteractive_ggtree}
\alias{create_noninteractive_ggtree}
\title{Create a non-interactive \code{ggtree} object for presenting mutation and lineage data}
\usage{
create_noninteractive_ggtree(
  ggtree_data,
  branch_col,
  lins,
  shapes,
  colours,
  colour_limits
)
}
\arguments{
\item{ggtree_data}{Tree data for passing to \code{ggtree}.}

\item{branch_col}{Scalar string. The name of a column within \code{ggtree_data} for which
the \code{ggtree} object should be produced.}

\item{lins}{List. Defines the virus lineages that are under study. Three entries are
present. \code{lins$lineages} a string vector of the lineages. \code{lins$nodes} an integer
vector of node numbers, these correspond to nodes in \code{ggtree_data$node}.
\code{lins$node_names} a string vector giving the name of the lineage. For \code{nodes} and
\code{node_names} the order of entries matches that for \code{lins}.}

\item{shapes}{Shapes for the branches and leaves in the tree.}

\item{colours}{Vector of colours.}

\item{colour_limits}{Min and max values for the colours.}
}
\value{
A \code{ggtree} object.
}
\description{
Create a non-interactive \code{ggtree} object for presenting mutation and lineage data
}
