% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_browser_data.R
\name{create_sequences_lookup}
\alias{create_sequences_lookup}
\title{Function to save a CSV file of all sequences for all clusterIDs}
\usage{
create_sequences_lookup(data_dir)
}
\arguments{
\item{data_dir}{The data directory for the application. Must have a `scanner_output`
subdirectory. Within `<data_dir>/scanner_output/` every subdirectory must contain a
`sequences.csv` file.}
}
\description{
The files `<data_dir>/scanner_output/*/sequences.csv` will be combined together to create the
output file `<data_dir>/sequences/all_sequences.csv`.
}
