% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{create_trees}
\alias{create_trees}
\title{Make tree views for the mutation / lineage data}
\usage{
create_trees(
  ggtree_data,
  branch_col,
  n_leaves,
  lins,
  sc0,
  cmuts,
  heatmap_width,
  heatmap_offset,
  heatmap_lab_offset,
  heatmap_fill = c(`FALSE` = "grey90", `TRUE` = "grey70"),
  mut_regex = NULL,
  colours = NULL,
  colour_limits = NULL
)
}
\arguments{
\item{ggtree_data}{Tree data for passing to \code{ggtree}.}

\item{branch_col}{Scalar string. A statistic for which the trees provide a summary. The name
of a column within \code{ggtree_data}.}

\item{n_leaves}{Scalar integer. The number of leaves in the tree.}

\item{lins}{List. Defines the virus lineages that are under study. Three entries are
present. \code{lins$lineages} a string vector of the lineages. \code{lins$nodes} an integer
vector of node numbers, these correspond to nodes in \code{ggtree_data$node}.
\code{lins$node_names} a string vector giving the name of the lineage. For \code{nodes} and
\code{node_names} the order of entries matches that for \code{lins}.}

\item{sc0, cmuts}{Data-frames.}

\item{heatmap_width, heatmap_offset}{Width relative to the tree and offset from the tree for the
heatmap.}

\item{heatmap_lab_offset}{Label-offset parameter for the constructed heatmap.}

\item{heatmap_fill}{Colours for filling the interior of the heatmap (which indicates the
presence / absence of a particular genotype). By default this is light grey for `FALSE` and
mid-grey for `TRUE`.}

\item{mut_regex}{Regular expression. Defines the mutations under study here.}

\item{colours}{Vector of colours.}

\item{colour_limits}{Min and max values for the colours.}
}
\value{
A list with several entries. Each entry is a \code{ggtree} object. The list names are
  "noninteractive", "with_interactivity_data", "with_heatmap", "interactive".
}
\description{
Make tree views for the mutation / lineage data
}
