% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_widget.R
\name{create_widget}
\alias{create_widget}
\title{Converts an interactive ggplot2 object into an HTML widget}
\usage{
create_widget(ggobj, width_svg, height_svg, sizing_width = 0.8)
}
\arguments{
\item{ggobj}{An interactive \code{ggplot2} object.}

\item{width_svg, height_svg}{Width and height for the output SVG.}

\item{sizing_width}{Sets the width option in \code{ggiraph::opts_sizing}.}
}
\value{
A \code{girafe} / \code{htmlwidget} object for use in a web browser.
}
\description{
Converts an interactive ggplot2 object into an HTML widget
}
