% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_browser_data.R
\name{empty_treeview}
\alias{empty_treeview}
\title{Function to create a treeview with all grey nodes to display}
\usage{
empty_treeview(treeview, treeview_dir, types = c("mutations", "sequences"))
}
\arguments{
\item{treeview}{RDS file containing an existing treeview plot}

\item{treeview_dir}{The directory where data should be read from / written to.}

\item{types}{Character vector of new variables to colour by}
}
\description{
Function to create a treeview with all grey nodes to display
}
