% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{extract_genotype_data}
\alias{extract_genotype_data}
\title{Extract the data about viral genotypes from a \code{ggtree} object}
\usage{
extract_genotype_data(ggobj, n_leaves, mut_regex)
}
\arguments{
\item{ggobj}{A \code{ggtree} object, as generated by \code{append_interactivity_data}. The
\code{data} entry for this object should contain the columns "node", "label" and a column for
each of the \code{mut_regex} values.}

\item{n_leaves}{Scalar integer. The number of leaves in the tree.}

\item{mut_regex}{String. Regular expression defining the mutations under study here. This
should be a subset of the column-names in \code{ggobj$data}.}
}
\value{
Data-frame.
}
\description{
Extract the data about viral genotypes from a \code{ggtree} object
}
