% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree_dataframe.R
\name{extract_tree_dataframe}
\alias{extract_tree_dataframe}
\title{Extracts node-specific data and annotates whether they are internal or a tip in the tree}
\usage{
extract_tree_dataframe(sc0, tr2, branch_cols)
}
\arguments{
\item{sc0}{data.frame.}

\item{tr2}{phylo.}

\item{branch_cols}{Character vector defining the statistics that must be extracted from
\code{sc0} for all nodes in the tree.}
}
\value{
data.frame. Containing a subset of the columns from \code{sc0} (including all
  \code{branch_cols}) and some additional columns. The additional columns are \code{lineages},
  \code{cocirc_summary}, \code{node}, \code{internal}, \code{lineages1}.
}
\description{
Extracts node-specific data and annotates whether they are internal or a tip in the tree
}
