% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_sina.R
\name{format_cluster_sina_data}
\alias{format_cluster_sina_data}
\title{Reformats data for a tree-plot for presentation in a sina plot}
\usage{
format_cluster_sina_data(pldf, varx, mut_regexp = NULL, lineage_regexp = NULL)
}
\arguments{
\item{pldf}{Data-frame. The data element of a tree plot.}

\item{varx}{String. Which variable in \code{pldf} should be plotted on the y-axis.}

\item{mut_regexp}{String. Regular-expression(s) for restricting the plot to a subset of the rows
in \code{pldf}. Only rows with an "allmuts" entry that matches one of these regular expressions
will be presented. If NULL, all rows are presented.}

\item{lineage_regexp}{String. Regular-expression(s) for restricting the lineages that are
presented in the plot. Only those rows of \code{pldf} with a "lineage" entry that matches one
of these regular expressions will be presented. If NULL, all rows are presented.}
}
\description{
Reformats data for a tree-plot for presentation in a sina plot
}
