% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_browser_data.R
\name{get_cluster_ID}
\alias{get_cluster_ID}
\title{Function to get node id from data_id column of ggplot
TO BE REMOVED AFTER TOOLTIPS TFPSCANNER PR IS MERGED}
\usage{
get_cluster_ID(tooltip_input)
}
\arguments{
\item{tooltip_input}{Character vector of tooltip content}
}
\description{
Function to get node id from data_id column of ggplot
TO BE REMOVED AFTER TOOLTIPS TFPSCANNER PR IS MERGED
}
