% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfpscanner.R
\name{get_clusternode_mlesky}
\alias{get_clusternode_mlesky}
\title{Run a fast treedater/mlesky analysis for a given node in the scanner output}
\usage{
get_clusternode_mlesky(
  u = 406318,
  scanner_env = readRDS("scanner-env-2021-03-03.rds")
)
}
\description{
Run a fast treedater/mlesky analysis for a given node in the scanner output
}
