% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_sina.R
\name{get_mutation_and_lineage}
\alias{get_mutation_and_lineage}
\title{Returns a string vector containing mutation and lineage information}
\usage{
get_mutation_and_lineage(x, mut_regexp = NULL, lineage_regexp = NULL)
}
\arguments{
\item{x}{A data-frame, it must contain a "allmuts" and a "lineage" column.}

\item{mut_regexp}{String. Regular-expression(s) for restricting the plot to a subset of the rows
in \code{pldf}. Only rows with an "allmuts" entry that matches one of these regular expressions
will be presented. If NULL, all rows are presented.}

\item{lineage_regexp}{String. Regular-expression(s) for restricting the lineages that are
presented in the plot. Only those rows of \code{pldf} with a "lineage" entry that matches one
of these regular expressions will be presented. If NULL, all rows are presented.}
}
\value{
Vector of strings, one for each row of \code{x}.
}
\description{
The strings are "_"-separated, with mutations on the left and lineages on the right.
In a given string, there is an entry for each regexp that matches.
The regex matches are "."-separated (there is n-1 dots on the LHS if there are n `mut_regexps`).
}
