% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sc0.R
\name{get_mutation_list}
\alias{get_mutation_list}
\title{Extract lists of 'defining' and 'all' mutations for each cluster in the data.frame `sc0`}
\usage{
get_mutation_list(sc0)
}
\arguments{
\item{sc0}{data.frame. Must contains columns \code{cluster_id}, \code{defining_mutations} and
\code{all_mutations}. The "mutation" columns contain a "|"-separated string of the mutations
present in any given cluster. Each row represents a cluster in the phylogeny.}
}
\value{
List of lists. Each entry in the named list corresponds to a cluster in the phylogeny.
  The inner lists all have entries "defining" and "all", character vectors defining the mutations
  that are present in the cluster.
}
\description{
Extract lists of 'defining' and 'all' mutations for each cluster in the data.frame `sc0`
}
