% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sc0.R
\name{get_mutation_tables}
\alias{get_mutation_tables}
\title{Extract data.frames containing the 'defining' and 'all' mutations for each cluster}
\usage{
get_mutation_tables(sc0)
}
\arguments{
\item{sc0}{data.frame. Must contains columns \code{cluster_id}, \code{defining_mutations} and
\code{all_mutations}. The "mutation" columns contain a "|"-separated string of the mutations
present in any given cluster. Each row represents a cluster in the phylogeny.}
}
\value{
List of two data.frames with names "all" and "defining". These contain all mutations-
  and just the defining mutations for each cluster in the phylogeny. The data.frames have
  identical structure with column names \code{cluster_id} and \code{mutation}.
}
\description{
Extract data.frames containing the 'defining' and 'all' mutations for each cluster
}
