% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_sina.R
\name{plot_cluster_sina}
\alias{plot_cluster_sina}
\title{Create a sina plot against lineage}
\usage{
plot_cluster_sina(
  pldf,
  varx = "logistic_growth_rate",
  mut_regexp = "S:A222V",
  lineage_regexp = NULL
)
}
\arguments{
\item{pldf}{Data-frame. The data element of a tree plot.}

\item{varx}{String. Which variable in \code{pldf} should be plotted on the y-axis.}

\item{mut_regexp}{String. Regular-expression(s) for restricting the plot to a subset of the rows
in \code{pldf}. Only rows with an "allmuts" entry that matches one of these regular expressions
will be presented. If NULL, all rows are presented.}

\item{lineage_regexp}{String. Regular-expression(s) for restricting the lineages that are
presented in the plot. Only those rows of \code{pldf} with a "lineage" entry that matches one
of these regular expressions will be presented. If NULL, all rows are presented.}
}
\value{
A \code{ggplot2} object storing the sina-cluster plot.
}
\description{
Create a sina plot against lineage
}
