% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_browser_data.R
\name{process_seq_table}
\alias{process_seq_table}
\title{Function to get lookup table of clusterID and sequence}
\usage{
process_seq_table(selected_folder, data_dir)
}
\arguments{
\item{selected_folder}{Folder name relating to a single clusterID}

\item{data_dir}{The data directory for the application. Must have a `scanner_output`
subdirectory.}
}
\description{
Function to get lookup table of clusterID and sequence
}
