% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_sina.R
\name{save_sina_plot}
\alias{save_sina_plot}
\title{Save a sina-cluster-plot to a file}
\usage{
save_sina_plot(
  ggobj,
  varx,
  output_dir,
  output_format = c("rds", "html"),
  width_svg = 8,
  height_svg = 8
)
}
\arguments{
\item{ggobj}{\code{ggplot2} object. Contains the plot that is to be saved.}

\item{varx}{Scalar string. Which variable is depicted in the plot?}

\item{output_dir}{File path. The directory where the plot will be stored.}

\item{output_format}{String (either \code{rds}, \code{html} or both). Default: both. In
which file format(s) should the plots be saved?}

\item{width_svg, height_svg}{The width and height of the plot (only used when
\code{output_format == "html"}).}
}
\value{
Invisibly returns the file paths (one for each output format) where the plots were
  saved.
}
\description{
Saves as either an htmlwidget (in a \code{html} file) or a ggplot object (in a \code{rds} file).
}
