% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{save_trees}
\alias{save_trees}
\title{Save the tree views to files}
\usage{
save_trees(
  tree_list,
  branch_col,
  output_dir,
  output_format = c("rds", "html"),
  include_date = FALSE,
  height_svg = NULL,
  width_svg = NULL
)
}
\arguments{
\item{tree_list}{List of \code{ggtree} objects. The "noninteractive" and "interactive"
entries will be saved to an external file. As generated by \code{create_trees}.}

\item{branch_col}{Scalar string. The name of the statistic that is under study here (e.g.,
"logistic_growth_rate" or "clock_outlier"). This name will be used in the output file paths.}

\item{output_dir}{Scalar string. The directory where the plots will be saved.}

\item{output_format}{String (either \code{rds} or \code{html} or both). Default: both. Which
file format(s) should the interactive tree plot(s) be placed? For \code{rds}, a \code{ggtree}
object will be placed in an \code{rds} file. For \code{html}, a \code{htmlwidget} will be
placed in a \code{html} file.}

\item{include_date}{Boolean. Should the file-paths include the current date?}

\item{height_svg, width_svg}{Scalar numeric. Height/width of the generated plots. Passed on to
`ggplot2::ggsave(..., height, width)` and `ggiraph::girafe(..., height_svg, width_svg)`.}
}
\value{
A named vector containing the file paths that were generated.
}
\description{
Save the tree views to files
}
