% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{sort_mutations}
\alias{sort_mutations}
\title{Sorts a vector of mutations}
\usage{
sort_mutations(muts)
}
\arguments{
\item{muts}{String. Vector of mutations. Each string must be a separate mutation
(e.g., "S:A243del"). The mutations have a prefix ("S:", "N:") and a positional description of
the protein-level mutation ("T205I" for Thr to Ile mutation at position 205).}
}
\value{
String. A vector of the same length as \code{muts}. The mutations are sorted by prefix
  and then by the location of the mutation.
}
\description{
Sorts a vector of mutations
}
