% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{create_noninteractive_ggtree}
\alias{create_noninteractive_ggtree}
\title{Create a non-interactive \code{ggtree} object for presenting mutation and lineage data}
\usage{
create_noninteractive_ggtree(
  ggtree_data,
  branch_col,
  lins,
  lin_nodes,
  lin_node_names,
  shapes,
  colours,
  colour_limits
)
}
\arguments{
\item{ggtree_data}{Tree data for passing to \code{ggtree}.}

\item{branch_col}{Scalar string. The name of a column within \code{ggtree_data} for which
the \code{ggtree} object should be produced.}

\item{lins}{String. Vector of lineages that are under study.}

\item{lin_nodes}{Integer. Vector of node numbers. The nodes are defined in
\code{ggtree_data$node}. The order of entries matches that for \code{lins}.}

\item{lin_node_names}{String. Name of the lineage. The order of entries matches that for
\code{lins}.}

\item{shapes}{Shapes for the branches and leaves in the tree.}

\item{colours}{Vector of colours.}

\item{colour_limits}{Min and max values for the colours.}
}
\value{
A \code{ggtree} object.
}
\description{
Create a non-interactive \code{ggtree} object for presenting mutation and lineage data
}
