% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{save_trees}
\alias{save_trees}
\title{Save the tree views to files}
\usage{
save_trees(tree_list, branch_col, n_leaves, output_dir, include_date = FALSE)
}
\arguments{
\item{tree_list}{List of \code{ggtree} objects. The "noninteractive" and "interactive"
entries will be saved to an external file. As generated by \code{create_trees}.}

\item{branch_col}{Scalar string. The name of the statistic that is under study here (e.g.,
"logistic_growth_rate" or "clock_outlier"). This name will be used in the output file paths.}

\item{n_leaves}{Scalar integer. The number of leaves in the tree.}

\item{output_dir}{Scalar string. The directory where the plots will be saved.}

\item{include_date}{Boolean. Should the file-paths include the current date?}
}
\value{
A named vector containing the file paths that were generated.
}
\description{
Save the tree views to files
}
