% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeview.R
\name{treeview}
\alias{treeview}
\title{Generate interactive tree visualisations and scatter plots for illustrating
scannint statistics.}
\usage{
treeview(
  e0,
  branch_cols = c("logistic_growth_rate", "clock_outlier"),
  mutations = c("S:A222V", "S:Y145H", "N:Q9L", "S:E484K"),
  lineages = c("AY\\\\.9", "AY\\\\.43", "AY\\\\.4\\\\.2"),
  output_dir = "treeview"
)
}
\arguments{
\item{e0}{Path to the scanner environment produced by \code{tfpscan}. Alternatively
can pass the environment directly.}

\item{branch_cols}{A character vector of statistics for which outputs should be
produced. The logistic growth rate plot will always be produced.}

\item{mutations}{A character vector of mutations which will be illustrated in a
heatmap}

\item{lineages}{A set of lineage names which will be used to subdivide outputs in
scatter plots.}

\item{output_dir}{Outputs will be saved in this directory. Will create the directory
if it does not exist.}
}
\value{
A ggtree plot
}
\description{
This will produce a set of html widgets which will highlight by colour and
tooltips statistics such as growth rate and molecular clock outliers.
}
